refactor(structure): rename desk
