+++ packages/sanity/src/plugins/customDocumentActions/shopifyDelete.tsx
@@ -78,7 +78,7 @@ export default (props: ShopifyDocumentActionProps): DocumentActionDescription | 
         try {
           await transaction.commit()
           // Navigate back to products root
-          router.navigateUrl({path: `${basePath}/desk/products`.replace(/\/+/, '/')})
+          router.navigateUrl({path: `${basePath}/structure/products`.replace(/\/+/, '/')})
         } catch (err) {
           let message = 'Unknown Error'
           if (err instanceof Error) message = err.message
@@ -118,7 +118,7 @@ export default (props: ShopifyDocumentActionProps): DocumentActionDescription | 
         try {
           await transaction.commit()
           // Navigate back to collections root
-          router.navigateUrl({path: `${basePath}/desk/collections`.replace(/\/+/, '/')})
+          router.navigateUrl({path: `${basePath}/structure/collections`.replace(/\/+/, '/')})
         } catch (err) {
           let message = 'Unknown Error'
           if (err instanceof Error) message = err.message
