+++ packages/sanity/src/structure/colorThemes.ts
