+++ packages/sanity/src/structure/people.ts
