+++ packages/sanity/src/structure/products.ts
