+++ packages/sanity/src/utils/defineStructure.ts
@@ -2,7 +2,7 @@ import {ConfigContext} from 'sanity'
 import {StructureBuilder} from 'sanity/structure'
 
 /**
- * Helper for creating and typing composable desk structure parts.
+ * Helper for creating and typing composable structure parts.
  */
 export default function defineStructure<StructureType>(
   factory: (S: StructureBuilder, context: ConfigContext) => StructureType,
